requirements.txt
gunicorn==20.0.4
Flask==1.1.1
googletrans==2.4.0
python-dotenv==0.12.0 